+++ commands/storage_deal.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 
 	"github.com/c2h5oh/datasize"
+	"github.com/filecoin-project/dealbot/lotus"
 	"github.com/filecoin-project/dealbot/tasks"
 	"github.com/urfave/cli/v2"
 )
@@ -50,7 +51,7 @@ var MakeStorageDealCmd = &cli.Command{
 }
 
 func makeStorageDeal(cctx *cli.Context) error {
-	clientConfig, node, closer, err := setupCLIClient(cctx)
+	nodeConfig, node, closer, err := lotus.SetupClientFromCLI(cctx)
 	if err != nil {
 		return err
 	}
@@ -80,7 +81,7 @@ func makeStorageDeal(cctx *cli.Context) error {
 		Verified:        verified,
 	}
 
-	return tasks.MakeStorageDeal(cctx.Context, clientConfig, node, task, func(msg string, keysAndValues ...interface{}) {
+	return tasks.MakeStorageDeal(cctx.Context, nodeConfig, node, task, func(msg string, keysAndValues ...interface{}) {
 		log.Infow(msg, keysAndValues...)
 	})
 }
